knoin_backend/requirements.txt
django
mysqlclient
djangorestframework
django-cors-headers
coreapi
djangorestframework-jwt
django-filter
pillow
docxtpl